swap_face_server/pyproject.toml
[tool.poetry]
name = "swap_face_server"
version = "0.1.0"
description = ""
authors = ["Walter <jinshuhaicc@gmail.com>"]
package-mode = false
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
numpy = "1.23.5"
opencv-python = "4.8.1.78"
onnx = "^1.17.0"
insightface = "^0.7.3"
psutil = "^6.1.0"
tk = "^0.1.0"
pillow = "^11.0.0"
customtkinter = "^5.2.2"
cv2-enumerate-cameras = "^1.1.17"
torch = "2.0.1"
torchvision = "0.15.2"
onnxruntime = "^1.16.3"
#opennsfw2 = "0.10.2"
protobuf = "4.23.2"
tqdm = "4.66.4"
gfpgan = "1.3.8"
tkinterdnd2 = "0.4.2"
pygrabber = "0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
